% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-local.R
\name{analyse-local_qda}
\alias{analyse-local_qda}
\alias{analyse_local.tbl_sensory_qda}
\title{Local analysis of sensory data}
\usage{
\method{analyse_local}{tbl_sensory_qda}(.data, ...)
}
\arguments{
\item{.data}{a sensory table}

\item{...}{not yet implemented}
}
\value{
a sensory table of local analysis
}
\description{
Perform local analysis on sensory table.
}
\examples{
data(perfume_qda_consumers)
(df <- specify(.data = perfume_qda_consumers, 
  panelist = consumer, 
  product = product, 
  attribute = intensity:green, 
  hedonic = NULL,
  method = "QDA"))
analyse_local(df)

# Using pipe \%>\%
data(perfume_qda_experts)
perfume_qda_experts \%>\% 
specify(
  panelist = panelist,
  product = product,
  session = session,
  pres_order = rank,
  attribute = spicy:wrapping,
  method = "QDA"
) \%>\% 
analyse_local()
}
