% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Sensory data analysis}
\usage{
analyse(tbl_sensory, choice, ...)
}
\arguments{
\item{tbl_sensory}{a sensory table}

\item{choice}{choice of analysis, see details.}

\item{...}{other arguments for specific method}
}
\description{
Perform statistical analysis on various sensory methods.
}
