% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-global.R
\name{analyse_global}
\alias{analyse_global}
\title{Global analysis of sensory data}
\usage{
analyse_global(.data, ...)
}
\arguments{
\item{.data}{a sensory table}

\item{...}{other arguments to pass on specific method}
}
\description{
Perform global analysis on sensory table.
}
