% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-liking.R
\name{analyse_liking}
\alias{analyse_liking}
\title{Liking analysis of sensory data}
\usage{
analyse_liking(data)
}
\arguments{
\item{data}{a sensory table}
}
\description{
Perform liking analysis on sensory table.
}
