% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-local.R
\name{analyse_local}
\alias{analyse_local}
\title{Local analysis of sensory data}
\usage{
analyse_local(.data, ...)
}
\arguments{
\item{.data}{a sensory table}

\item{...}{other arguments to pass on specific method}
}
\description{
Perform local analysis on sensory table.
}
