% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-penalty.R
\name{analyse_penalty}
\alias{analyse_penalty}
\title{Penalty analysis of sensory data}
\usage{
analyse_penalty(data, reference_value)
}
\arguments{
\item{data}{a sensory table}

\item{reference_value}{a score used as JAR value (reference)}
}
\description{
Perform penalty analysis on sensory table.
}
\examples{
perfume_jar \%>\%
  specify(
    sensory_method = "JAR",
    panelist = consumer,
    product = product,
    attribute = intensity:green,
    hedonic = liking
  ) \%>\%
  analyse_penalty(reference_value = 0)
}
