% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-performance.R
\name{analyse_performance}
\alias{analyse_performance}
\title{Global analysis of sensory data}
\usage{
analyse_performance(data)
}
\arguments{
\item{data}{a sensory table}
}
\description{
Perform global analysis on sensory table.
}
\examples{
sensochoc \%>\%
  specify(
    sensory_method = "QDA",
    panelist = panelist,
    product = product,
    session = session,
    pres_order = rank,
    attribute = cocoa_a:granular
  ) \%>\%
  analyse_performance()
}
