% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-preference.R
\name{analyse_preference}
\alias{analyse_preference}
\title{Preference analysis of sensory data}
\usage{
analyse_preference(data)
}
\arguments{
\item{data}{a sensory table}
}
\description{
Perform preference analysis on sensory table.
}
