% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status.R
\name{check_status}
\alias{check_status}
\title{Check status for CATA JAR}
\usage{
check_status(x, ref)
}
\arguments{
\item{x}{score for product}

\item{ref}{score for ideal product}
}
\value{
a vector
}
\description{
Check status for CATA JAR
}
