% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_template}
\alias{create_template}
\title{Create Data Tabulation Template}
\usage{
create_template(.data, panelist, attribute)
}
\arguments{
\item{.data}{output from \code{generate_design} function}

\item{panelist}{number of panelist}

\item{attribute}{a numeric value of number or sensory attributes or vector of sensory terms (lexicon)}
}
\value{
Returns a dataframe of sensory design with Panelist, Order, Product, and Sensory Attributes columns
}
\description{
Create template for inputting data collected from sensory research
}
\examples{

design <- generate_design(n_panelist = 20, product = 5, use_coding = TRUE)
design

create_template(design, panelist = Panelist, attribute = c("Sweetness", "Mint", "Green", "Herbal"))


}
