% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template_.R
\name{create_template_}
\alias{create_template_}
\title{Create template for data input}
\usage{
create_template_(.data, panelist, attribute)
}
\arguments{
\item{.data}{output from \code{generate_design} function}

\item{panelist}{column containing panelist information}

\item{attribute}{a numeric value of number or sensory attributes or vector of sensory terms (lexicon)}
}
\value{
Returns a dataframe of sensory design with Panelist, Order, Product, and Sensory Attributes columns
}
\description{
Create template for inputting data collected from sensory research
}
\examples{

design <- generate_design_(n_panelist = 20, product = 5, use_coding = TRUE)
design

create_template_(design, panelist = Panelis, attribute = c("Sweetness", "Mint", "Green", "Herbal"))


}
