% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminate.R
\name{discriminate}
\alias{discriminate}
\title{Discrimination test}
\usage{
discriminate(n_correct, n_panelist, method, conf_level = 0.95)
}
\arguments{
\item{n_correct}{the number of correct answer}

\item{n_panelist}{the total number of panelist}

\item{method}{the discimination test protocol. One of 2-AFC, 3-AFC, Duo-Trio, Triangle, Tetrad, Hexad, and Two-out-of-Five.}

\item{conf_level}{confidence level for the test}
}
\description{
Perform statistical test for sensory discrimination method.
}
\examples{
discriminate(n_correct = 30, n_panelist = 80, method = "Triangle")
}
