% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ca.R
\name{do_ca}
\alias{do_ca}
\title{Perform CA on dataframe}
\usage{
do_ca(.data, product, liking = NULL)
}
\arguments{
\item{.data}{a dataframe or tibble.}

\item{product}{name of column containing product information.}

\item{liking}{name of column containing hedonic liking (default to NULL).}
}
\value{
Returns a list containing information of CA results.
}
\description{
Pipe friendly CA
}
