% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_cata_penalty.R
\name{do_cata_penalty}
\alias{do_cata_penalty}
\title{Perform JAR on CATA with Ideal data}
\usage{
do_cata_penalty(.data, panelist, product, ideal_product, attribute, liking)
}
\arguments{
\item{.data}{a dataframe}

\item{panelist}{a column containing information of panelist}

\item{product}{a column containing information of product}

\item{ideal_product}{name of ideal product}

\item{attribute}{set of columns containing information of  sensory attributes}

\item{liking}{a column containing information of liking}
}
\value{
a tibble
}
\description{
Perform JAR on CATA with Ideal data
}
