% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_jar.R
\name{do_jar}
\alias{do_jar}
\title{Perform JAR analysis}
\usage{
do_jar(.data, attribute, jar_value, liking)
}
\arguments{
\item{.data}{a dataframe}

\item{attribute}{set of columns containing information of  sensory attributes}

\item{jar_value}{JAR value (numeric)}

\item{liking}{a column containing information of liking}
}
\value{
a tibble
}
\description{
Perform JAR analysis
}
