% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_design.R
\name{generate_design}
\alias{generate_design}
\title{Generate Design}
\usage{
generate_design(n_panelist, product, use_coding = TRUE)
}
\arguments{
\item{n_panelist}{number of panelist}

\item{product}{a numeric value of number or product or vector of product names}

\item{use_coding}{wheteher to generate random three digit number for labeling}
}
\value{
Returns a dataframe of sensory design with Panelist column and Order of product presentation columns
}
\description{
Generate design of presentation order based on Williams design.
}
\examples{
generate_design(n_panelist = 10, product = letters[1:14], use_coding = FALSE)
generate_design(n_panelist = 10, product = letters[1:14], use_coding = TRUE)

design <- generate_design(n_panelist = 20, product = 5, use_coding = TRUE)
design


}
