% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eig.R
\name{get_eig}
\alias{get_eig}
\title{Get eigenvalues}
\usage{
get_eig(.res)
}
\arguments{
\item{.res}{Output from multivariate analyses}
}
\value{
a tibble
}
\description{
Function to augment eigenvalues from multivariate analysis
}
