% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_summary.R
\name{get_global_summary}
\alias{get_global_summary}
\title{Get summary for global analysis}
\usage{
get_global_summary(.res, choice, axes = c(1, 2))
}
\arguments{
\item{.res}{output of PCA or CA}

\item{choice}{choices: "product" or "attribute"}

\item{axes}{dimension of interest}
}
\value{
a tibble
}
\description{
Get summary for global analysis
}
