% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance.R
\name{get_performance}
\alias{get_performance}
\title{Get performance of sensory panel and panelist}
\usage{
get_performance(.data, panelist = NULL, product = NULL,
  session = NULL, pres_order = NULL, attribute = NULL)
}
\arguments{
\item{.data}{a dataframe}

\item{panelist}{column contaning panelist information}

\item{product}{column contaning product information}

\item{session}{column containing session information}

\item{pres_order}{column contaning presentation order information}

\item{attribute}{vector of columns of sensory attributes}
}
\value{
a named list of four dataframe
}
\description{
Perform calculation to get performance metrics
}
\examples{
\dontrun{
library(SensoMineR)
data(chocolates)
attrs <- colnames(sensochoc)[-c(1:4)]
out <-
  get_performance(sensochoc, panelist = Panelist, product = Product, session = Session, pres_order = Rank, attribute = attrs)
out
}

}
