% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats_.R
\name{get_stats_}
\alias{get_stats_}
\title{Get statistics}
\usage{
get_stats_(.data, stat = "anova")
}
\arguments{
\item{.data}{a dataframe}

\item{stat}{statistical procedure to be performed. Choice: "anova", "cochran", and "friedman"}
}
\value{
a tibble
}
\description{
Perform univariate statistical analysis for all sensory attributes simultaniously
}
