% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_attribute.R
\name{glance_attribute}
\alias{glance_attribute}
\title{Glance sensory attribute}
\usage{
glance_attribute(res_global, dimension = c(1, 2))
}
\arguments{
\item{res_global}{output of global analysis}

\item{dimension}{dimension to focus, integer vector of length 2}
}
\value{
a dataframe
}
\description{
Evaluate sensory attribute in global analysis.
}
