% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance-eigenvalue.R
\name{glance_eigenvalue}
\alias{glance_eigenvalue}
\title{Glance eigenvalues}
\usage{
glance_eigenvalue(res_global)
}
\arguments{
\item{res_global}{output of global analysis}
}
\value{
a dataframe
}
\description{
Evaluate the dimension of sensory space in global analysis.
}
