% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance-panel.R
\name{glance_panel}
\alias{glance_panel}
\title{Glance panel}
\usage{
glance_panel(res_performance)
}
\arguments{
\item{res_performance}{output of performance analysis}
}
\description{
Evaluate panel performance.
}
