% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance-panelist.R
\name{glance_panelist}
\alias{glance_panelist}
\title{Glance panelist}
\usage{
glance_panelist(res_performance, metric)
}
\arguments{
\item{res_performance}{output of performance analysis}

\item{metric}{discrimination, agreement, consistency}
}
\description{
Evaluate panelist performance.
}
