% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance-product.R
\name{glance_product}
\alias{glance_product}
\title{Glance product}
\usage{
glance_product(res_global, dimension = c(1, 2))
}
\arguments{
\item{res_global}{output of global analysis}

\item{dimension}{dimension to focus, integer vector of length 2}
}
\description{
Evaluate product in global analysis.
}
