% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect-attribute.R
\name{inspect_attribute}
\alias{inspect_attribute}
\title{Inspect sensory attribute}
\usage{
inspect_attribute(res_global, dimension = c(1, 2))
}
\arguments{
\item{res_global}{output of global analysis}

\item{dimension}{dimension to focus, integer vector of length 2}
}
\description{
Evaluate sensory attribute in global analysis.
}
