% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect-panel.R
\name{inspect_panel}
\alias{inspect_panel}
\title{Inspect panel}
\usage{
inspect_panel(res_performance)
}
\arguments{
\item{res_performance}{output of performance analysis}
}
\description{
Evaluate panel performance.
}
