% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect-panelist.R
\name{inspect_panelist}
\alias{inspect_panelist}
\title{Inspect panelist}
\usage{
inspect_panelist(res_performance, metric)
}
\arguments{
\item{res_performance}{output of performance analysis}

\item{metric}{discrimination, agreement, consistency}
}
\description{
Evaluate panelist performance.
}
