% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect-panelist-liking.R
\name{inspect_panelist_liking}
\alias{inspect_panelist_liking}
\title{Inspect liking}
\usage{
inspect_panelist_liking(res_liking, dimension = c(1, 2))
}
\arguments{
\item{res_liking}{output of preference analysis}

\item{dimension}{dimension to focus, integer vector of length 2}
}
\description{
Evaluate panelist in liking analysis.
}
