% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect-preference-panelist.R
\name{inspect_preference_panelist}
\alias{inspect_preference_panelist}
\title{Inspect preference}
\usage{
inspect_preference_panelist(res_preference, dimension = c(1, 2))
}
\arguments{
\item{res_preference}{output of preference analysis}

\item{dimension}{dimension to focus, integer vector of length 2}
}
\description{
Evaluate panelist in preference analysis.
}
