% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect-product.R
\name{inspect_product}
\alias{inspect_product}
\title{Inspect product}
\usage{
inspect_product(res_global, dimension = c(1, 2))
}
\arguments{
\item{res_global}{output of global analysis}

\item{dimension}{dimension to focus, integer vector of length 2}
}
\description{
Evaluate product in global analysis.
}
