% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect-product-preference.R
\name{inspect_product_preference}
\alias{inspect_product_preference}
\title{Inspect preference}
\usage{
inspect_product_preference(res_preference, dimension = c(1, 2))
}
\arguments{
\item{res_preference}{output of preference analysis}

\item{dimension}{dimension to focus, integer vector of length 2}
}
\description{
Evaluate product in preference analysis.
}
