% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect-space.R
\name{inspect_space}
\alias{inspect_space}
\title{Inspect sensory space}
\usage{
inspect_space(res_global)
}
\arguments{
\item{res_global}{output of global analysis}
}
\value{
a dataframe
}
\description{
Evaluate the dimension of sensory space in global analysis.
}
