% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{meta_info}
\alias{meta_info}
\title{Title}
\usage{
meta_info(x, meta = c("panelist", "product", "attribute", "hedonic"))
}
\arguments{
\item{x}{a dataframe of class `tbl_sensory`}

\item{meta}{metadata, one of "panelist", "product", "attribute", or "hedonic"}
}
\value{
a metadata
}
\description{
Title
}
