% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_data.R
\name{nest_data}
\alias{nest_data}
\title{Nest sensory dataset}
\usage{
nest_data(.data, panelist, product, pres_order, session = NULL)
}
\arguments{
\item{.data}{a dataframe}

\item{panelist}{column contaning panelist information}

\item{product}{column contaning product information}

\item{pres_order}{column contaning presentation order information}

\item{session}{column contaning replicate/session information}
}
\value{
a tibble
}
\description{
Function to nest dataset to its sensory attribute
}
