% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-anova.R
\name{perform_anova}
\alias{perform_anova}
\title{Analysis of variance}
\usage{
perform_anova(tbl_sensory)
}
\arguments{
\item{tbl_sensory}{a sensory table}
}
\description{
Perform Analysis of variance (ANOVA) on sensory table.
}
