% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-ca.R
\name{perform_ca}
\alias{perform_ca}
\title{Correspondance Analysis}
\usage{
perform_ca(tbl_sensory)
}
\arguments{
\item{tbl_sensory}{a sensory table}
}
\description{
Perform Correspondance Component Analysis (CA) on sensory table.
}
