% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-cochran.R
\name{perform_cochran}
\alias{perform_cochran}
\title{Cochran's Q test}
\usage{
perform_cochran(tbl_sensory)
}
\arguments{
\item{tbl_sensory}{a sensory table}
}
\description{
Perform Cochran's Q test on sensory table.
}
