% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-friedman.R
\name{perform_friedman}
\alias{perform_friedman}
\title{Friedman rank sum test}
\usage{
perform_friedman(tbl_sensory)
}
\arguments{
\item{tbl_sensory}{a sensory table}
}
\description{
Perform Friedman rank sum test on sensory table.
}
