% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-liking-analysis.R
\name{perform_liking_analysis}
\alias{perform_liking_analysis}
\title{Liking analysis}
\usage{
perform_liking_analysis(tbl_sensory)
}
\arguments{
\item{tbl_sensory}{a sensory table}
}
\description{
Perform liking analysis on hedonic score or rating.
}
