% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-mdpref.R
\name{perform_mdpref}
\alias{perform_mdpref}
\title{Internal preference mapping}
\usage{
perform_mdpref(tbl_sensory)
}
\arguments{
\item{tbl_sensory}{a sensory table}
}
\description{
Perform internal preference mapping (MDPref) on sensory table.
}
