% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-panel-analysis.R
\name{perform_panel_analysis}
\alias{perform_panel_analysis}
\title{Panel analysis}
\usage{
perform_panel_analysis(tbl_sensory)
}
\arguments{
\item{tbl_sensory}{a sensory table}
}
\description{
Perform analysis on panel and panelist's performance.
}
