% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-pca.R
\name{perform_pca}
\alias{perform_pca}
\title{Principal Component Analysis}
\usage{
perform_pca(tbl_sensory)
}
\arguments{
\item{tbl_sensory}{a sensory table}
}
\description{
Perform Principal Component Analysis (PCA) on sensory table.
}
