% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-penalty.R
\name{perform_penalty}
\alias{perform_penalty}
\title{Penalty analysis}
\usage{
perform_penalty(tbl_sensory, ref_value)
}
\arguments{
\item{tbl_sensory}{a sensory table}

\item{ref_value}{a score used as JAR reference value}
}
\description{
Perform penalty analysis on sensory table.
}
