% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform-prefmap.R
\name{perform_prefmap}
\alias{perform_prefmap}
\title{External preference mapping}
\usage{
perform_prefmap(tbl_sensory)
}
\arguments{
\item{tbl_sensory}{a sensory table}
}
\description{
Perform external preference mapping (PrefMap) on sensory table.
}
