% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{perfume_jar}
\alias{perfume_jar}
\title{Perfume QDA Consumers}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1236 rows and 24 columns.}
\source{
Analyzing Sensory Data with R \url{www.sensorywithr.org}
}
\usage{
perfume_jar
}
\description{
This dataset containing information about 12 perfumes that evaluated by 103 consumer panelists using predefined lexicons
}
\details{
\describe{
  \item{consumer}{consumers' id}
  \item{product}{perfumes}
  \item{freshness-green}{lexicons regarding perfume, using JAR scale with 0 as JAR value}
  \item{liking}{hedonic liking score}
}
}
\keyword{datasets}
