% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{perfume_qda_consumers}
\alias{perfume_qda_consumers}
\title{Perfume QDA Consumers}
\format{Dataframe dengan 142 baris dan 5 kolom

\describe{
  \item{consumer}{consumers' id}
  \item{product}{perfumes}
  \item{intensity-green}{lexicons regarding perfume}
}}
\source{
Analyzing Sensory Data with R \url{www.sensorywithr.org}
}
\usage{
perfume_qda_consumers
}
\description{
This dataset containing information about 12 perfumes that evaluated by 103 consumer panelists using predefined lexicons
}
\keyword{datasets}
