% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{perfume_qda_experts}
\alias{perfume_qda_experts}
\title{Perfume QDA Experts}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 288 rows and 16 columns.}
\source{
Analyzing Sensory Data with R \url{www.sensorywithr.org}
}
\usage{
perfume_qda_experts
}
\description{
This dataset containing information about 12 perfumes that evaluated by 12 expert panelists using predefined lexicons
}
\details{
\describe{
  \item{panelist}{panelists' id}
  \item{product}{perfumes}
  \item{session}{code of sensory evalutation session}
  \item{rank}{presentation order}
  \item{spicy-wrapping}{lexicons regarding perfume}
}
}
\keyword{datasets}
