% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_attribute.R
\name{plot_attribute}
\alias{plot_attribute}
\title{Plot attribute}
\usage{
plot_attribute(.res, axes = c(1, 2), main = "", lab_size = 4)
}
\arguments{
\item{.res}{Output from multivariate analyses}

\item{axes}{Selection of dimentions}

\item{main}{Title of the plot}

\item{lab_size}{Label size}
}
\value{
a plot
}
\description{
Plot attribute on sensory space
}
