% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eigen.R
\name{plot_eigen}
\alias{plot_eigen}
\title{Plot Screeplot}
\usage{
plot_eigen(res, threshold = TRUE)
}
\arguments{
\item{res}{output of PCA, CA, MCA, MFA}

\item{threshold}{whether to draw threshold eigenvalue line}
}
\value{
ggplot2 object
}
\description{
Create screeplot
}
