% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jar.R
\name{plot_jar}
\alias{plot_jar}
\title{Plot JAR}
\usage{
plot_jar(res, frequency_threshold = 20, drop_threshold = 1,
  title = "Penalty analysis", xlab = "Frequency (\%)",
  ylab = "Liking drop")
}
\arguments{
\item{res}{output of JAR analysis}

\item{frequency_threshold}{threshold for frequency}

\item{drop_threshold}{threshold for liking drop}

\item{title}{title}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
ggplot2 object
}
\description{
Plot JAR
}
