% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{plot_radar}
\alias{plot_radar}
\title{Radar Plot}
\usage{
plot_radar(.data, scale_min = 1, scale_max = 10, point_size = 4,
  line_width = 1, title = "", subtitle = "", caption = "")
}
\arguments{
\item{.data}{an aggregated dataset containing columns of product and sensory attributes}

\item{scale_min}{minimum value of the scales}

\item{scale_max}{maximum value of the scales}

\item{point_size}{point size}

\item{line_width}{line width}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}

\item{caption}{caption of the plot}
}
\value{
a radar plot
}
\description{
Create radar plot of sensory attributes for each product
}
