% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare sensory table}
\usage{
prepare(.data, panelist = NULL, product = NULL, pres_order = NULL,
  attribute = NULL, hedonic = NULL)
}
\arguments{
\item{.data}{a dataframe}

\item{panelist}{panelist column}

\item{product}{product column}

\item{pres_order}{presentation order column}

\item{attribute}{sensory attribute columns}

\item{hedonic}{hedonic column}
}
\value{
a sensory table (dataframe with class of \code{tbl_sensory})
}
\description{
Prepare a raw dataframe with minimal sensory informations (panelist, product, and attribute) into a sensory table fur further processing.
}
\examples{
data(perfume_qda_consumer)
(df <- prepare(.data = perfume_qda_consumers, 
  panelist = consumer, 
  product = product, 
  attribute = intensity:green, 
  hedonic = NULL))
}
