% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cata.R
\name{prepare_cata}
\alias{prepare_cata}
\title{Prepare data for CATA analysis}
\usage{
prepare_cata(.data, panelist, product, attribute, liking)
}
\arguments{
\item{.data}{a dataframe}

\item{panelist}{name of column containing panelist information}

\item{product}{name of column containing product information}

\item{attribute}{set of columns containing information of  sensory attributes}

\item{liking}{name of column containing liking information}
}
\value{
@ a tibble
}
\description{
Prepare data for CATA analysis
}
