% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_panel.R
\name{see_panel}
\alias{see_panel}
\title{See performance of whole panel}
\usage{
see_panel(res)
}
\arguments{
\item{res}{output from get_performance}
}
\value{
a dataframe containing p.value from ANOVA model
}
\description{
See the performance of panel as a whole
}
