% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_panelist.R
\name{see_panelist}
\alias{see_panelist}
\title{See Panelist}
\usage{
see_panelist(res, ability = c("discrimination", "consensus",
  "consistency"))
}
\arguments{
\item{res}{output from get_performance}

\item{ability}{character vector specifying panelist ability of interest, the choices are "discrimination", "consensus", and "consistency"}
}
\value{
a dataframe containing p-values of ANOVA model II (discrimination), or correlation values between Product coefficent of ANOVA model I and model II (consensus), or RMSE (consistency)
}
\description{
See the performance of each panelist in a panel
}
