% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensehubr-package.R
\docType{package}
\name{sensehubr-package}
\alias{sensehubr}
\alias{sensehubr-package}
\title{sensehubr: A Tidy Approach for Sensory Analyses}
\description{
sensehubr aims to provide set of tools for
    performing sensory analyses using tidy phylosophy. This package wraps
    around some core function from sensR, BradleyTerry2, FactoMineR, and
    SensoMineR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aswansyahputra/sensehubr}
  \item Report bugs at \url{https://github.com/aswansyahputra/sensehubr/issues}
}

}
\author{
\strong{Maintainer}: Muhammad Aswan Syahputra \email{aswansyahputra@sensolution.id}

}
\keyword{internal}
