% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{sensochoc}
\alias{sensochoc}
\title{Sensory data for 6 chocolates}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 348 rows and 18 columns.
}
\source{
SensoMineR package
}
\usage{
sensochoc
}
\description{
The data used here refer to the sensory description of six varieties of chocolates sold in France: each chocolate was evaluated twice by 29 panelists according to 14 sensory descriptors.
}
\details{
\describe{
  \item{panelist}{panelists' id}
  \item{session}{code of sensory evalutation session}
  \item{rank}{presentation order}
  \item{product}{chocolates}
  \item{cocoa_a - granular}{lexicons regarding chocolate}
}
}
\keyword{datasets}
