% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.R
\name{specify}
\alias{specify}
\title{Specify sensory table}
\usage{
specify(data, sensory_method = c("QDA", "CATA", "RATA", "FCP", "FP",
  "JAR", "IPM"), panelist, product, session = NULL, pres_order = NULL,
  attribute, hedonic = NULL)
}
\arguments{
\item{data}{a dataframe}

\item{sensory_method}{method of sensory evaluation, available methods are QDA, CATA, RATA, FCP, FP, JAR, IPM}

\item{panelist}{panelist column}

\item{product}{product column}

\item{session}{session column}

\item{pres_order}{presentation order column}

\item{attribute}{sensory attribute columns}

\item{hedonic}{hedonic column}
}
\value{
a sensory table (dataframe with class of \code{tbl_sensory})
}
\description{
Specify sensory informations into a raw dataframe. The minimal sensory informations are the panelist, the product, the sensory attributes, and the method in which the evaluation was conducted. Additonally the session, the presentation order and hedonic evaluation can also be specified.
}
\examples{
(df <- specify(
  data = perfume_qda_consumers,
  sensory_method = "QDA",
  panelist = consumer,
  product = product,
  attribute = intensity:green,
  hedonic = NULL
))

perfume_qda_experts \%>\%
  specify(
    sensory_method = "QDA",
    panelist = panelist,
    product = product,
    session = session,
    pres_order = rank,
    attribute = spicy:wrapping
  )
}
