% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-sum.R
\name{tbl_sum.tbl_sensory}
\alias{tbl_sum.tbl_sensory}
\title{Create metadata for sensory table}
\usage{
\method{tbl_sum}{tbl_sensory}(x)
}
\arguments{
\item{x}{object of interest}
}
\value{
a dataframe with class of `tbl_sensory`
}
\description{
Create metadata for sensory table
}
