% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise-global}
\alias{visualise-global}
\alias{visualise.tbl_sensory_global}
\title{Visualise sensory space}
\usage{
\method{visualise}{tbl_sensory_global}(res, choice = c("product",
  "attribute", "eigenvalue"), dimension = c(1, 2), repel = FALSE,
  colour_by = c("none", "quality", "contribution"), title = "default",
  ...)
}
\arguments{
\item{res}{output of global analysis}

\item{choice}{component of interest, valid values are "product" and "attribute"}

\item{dimension}{dimension to focus, integer vector of length 2}

\item{repel}{avoid overplotting of text label}

\item{colour_by}{colour point or line based on specific properties, valid values are "quality" and "contribution"}

\item{title}{a title to use in plot}

\item{...}{not yet implemented}
}
\value{
a ggplot object
}
\description{
Plot product representations or attributes on sensory spaces.
}
\examples{
perfume_qda_experts \%>\%
  specify(
    sensory_method = "QDA",
    panelist = panelist,
    product = product,
    session = session,
    pres_order = rank,
    attribute = spicy:wrapping
  ) \%>\%
  analyse(choice = "global") \%>\%
  visualise(choice = "attribute", colour_by = "contribution")
}
