% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise-local}
\alias{visualise-local}
\alias{visualise.tbl_sensory_local}
\title{Visualise sensory properties}
\usage{
\method{visualise}{tbl_sensory_local}(res, min_scales = 0,
  max_scales = 10, point_size = 4, line_width = 1,
  title = "Radar plot of sensory properties",
  legend_position = "bottom", ...)
}
\arguments{
\item{res}{result of local or global analysis}

\item{min_scales}{minimum value of the scales}

\item{max_scales}{maximum value of the scales}

\item{point_size}{point size}

\item{line_width}{line width}

\item{title}{a title to use in plot}

\item{legend_position}{position of legend, valid values are "top", "right", "bottom", and "left"}

\item{...}{not yet implemented}
}
\value{
a ggplot object
}
\description{
Create representation of sensory properties of each product in form of radar plot.
}
\examples{
perfume_qda_experts \%>\%
  specify(
    sensory_method = "QDA",
    panelist = panelist,
    product = product,
    session = session,
    pres_order = rank,
    attribute = spicy:wrapping
  ) \%>\%
  analyse_local() \%>\%
  visualise()
}
