% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise}
\alias{visualise}
\title{Visualise product or sensory attributes}
\usage{
visualise(res, ...)
}
\arguments{
\item{res}{result of local or global analysis}

\item{...}{other arguments to pass on specific visualisation method}
}
\description{
Make comparison between products and sensory attributes in visual representation.
}
