% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise.tbl_sensory_discrim}
\alias{visualise.tbl_sensory_discrim}
\title{Visualise Perceptual distribution}
\usage{
\method{visualise}{tbl_sensory_discrim}(res,
  title = "Perceptual distribution", ...)
}
\arguments{
\item{res}{result of disrimination test}

\item{title}{a title to use in plot}

\item{...}{not yet implemented}
}
\description{
Plot perceptual distribution obtained from discimination test.
}
