% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualise.tbl_sensory_prefmap}
\alias{visualise.tbl_sensory_prefmap}
\title{Visualise External Preference Map}
\usage{
\method{visualise}{tbl_sensory_prefmap}(
  res,
  title = "External Preference Mapping",
  xlab = "Dim 1",
  ylab = "Dim 2",
  ...
)
}
\arguments{
\item{res}{result of external preference mapping}

\item{title}{a title to use in plot}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{...}{not yet implemented}
}
\description{
Plot preference map of products in sensory space.
}
