% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-global.R
\name{analyse_global}
\alias{analyse_global}
\title{Global analysis of sensory data}
\usage{
analyse_global(.data, ...)
}
\arguments{
\item{.data}{a sensory table}

\item{...}{other arguments to pass on specific method}
}
\description{
Perform global analysis on sensory table.
}
\examples{
data(perfume_qda_consumers)
(df <- specify(.data = perfume_qda_consumers, 
  panelist = consumer, 
  product = product, 
  attribute = intensity:green, 
  hedonic = NULL,
  method = "QDA"))
analyse_global(df)

# Using pipe \%>\%
data(perfume_qda_experts)
perfume_qda_experts \%>\% 
specify(
  panelist = panelist,
  product = product,
  session = session,
  pres_order = rank,
  attribute = spicy:wrapping,
  method = "QDA"
) \%>\% 
analyse_global()

}
