% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_global.R
\name{analyse_global}
\alias{analyse_global}
\alias{analyse_global.tbl_sensory_qda}
\title{Global analysis for sensory data}
\usage{
analyse_global(.data)

\method{analyse_global}{tbl_sensory_qda}(.data)
}
\arguments{
\item{.data}{a sensory table}
}
\value{
a sensory table of global analysis
}
\description{
Perform global analysis on sensory table.
}
\examples{
data(perfume_qda_consumers)
(df <- specify(.data = perfume_qda_consumers, 
  panelist = consumer, 
  product = product, 
  attribute = intensity:green, 
  hedonic = NULL,
  method = "QDA"))
analyse_global(df)

# Using pipe \%>\%
data(perfume_qda_experts)
perfume_qda_experts \%>\% 
specify(
  panelist = panelist,
  product = product,
  session = session,
  pres_order = rank,
  attribute = spicy:wrapping,
  method = "QDA"
) \%>\% 
analyse_global()
}
