% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-global.R
\name{analyse_global}
\alias{analyse_global}
\title{Global analysis of sensory data}
\usage{
analyse_global(data, ...)
}
\arguments{
\item{data}{a sensory table}

\item{...}{other arguments to pass on specific method}
}
\description{
Perform global analysis on sensory table.
}
\examples{
(df <- specify(
  data = perfume_qda_consumers,
  sensory_method = "QDA",
  panelist = consumer,
  product = product,
  attribute = intensity:green,
  hedonic = NULL
))
analyse_global(df)

# Using pipe \%>\%
perfume_qda_experts \%>\%
  specify(
    sensory_method = "QDA",
    panelist = panelist,
    product = product,
    session = session,
    pres_order = rank,
    attribute = spicy:wrapping
  ) \%>\%
  analyse_global()
}
