% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_local.R
\name{analyse_local}
\alias{analyse_local}
\title{Local analysis of sensory data}
\usage{
analyse_local(.data, method = "QDA")
}
\arguments{
\item{.data}{a sensory table}

\item{method}{sensory method, one of "QDA", "CATA", ....}
}
\value{
a sensory table
}
\description{
Perform local analysis on sensory table.
}
\examples{
data(perfume_qda_consumers)
(df <- prepare(.data = perfume_qda_consumers, 
  panelist = consumer, 
  product = product, 
  attribute = intensity:green, 
  hedonic = NULL))
analyse_local(df, method = "QDA")

# Using pipe \%>\%
perfume_qda_consumers \%>\% 
prepare(panelist = consumer, 
  product = product, 
  attribute = intensity:green, 
  hedonic = NULL) \%>\% 
analyse_local(method = "QDA")
}
