% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-local.R
\name{analyse_local}
\alias{analyse_local}
\title{Local analysis of sensory data}
\usage{
analyse_local(data, ...)
}
\arguments{
\item{data}{a sensory table}

\item{...}{other arguments to pass on specific method}
}
\description{
Perform local analysis on sensory table.
}
\examples{
(df <- specify(
  data = perfume_qda_consumers,
  sensory_method = "QDA",
  panelist = consumer,
  product = product,
  attribute = intensity:green,
  hedonic = NULL
))
analyse_local(df)

# Using pipe \%>\%
perfume_qda_experts \%>\%
  specify(
    sensory_method = "QDA",
    panelist = panelist,
    product = product,
    session = session,
    pres_order = rank,
    attribute = spicy:wrapping
  ) \%>\%
  analyse_local()
}
