% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-template.R
\name{create_template}
\alias{create_template}
\title{Create template for sensory data}
\usage{
create_template(data, attribute, hedonic = FALSE)
}
\arguments{
\item{data}{output from \code{prepare} function}

\item{attribute}{a numeric value of number or sensory attributes or vector of sensory terms (lexicon)}

\item{hedonic}{whether to include hedonic rating or not}
}
\value{
Returns a dataframe of sensory design with panelist, presentation order, product, blind code (optional) and sensory attributes columns
}
\description{
Create template for inputting data that collected from sensory research.
}
\examples{
design <- prepare(n_panelist = 20, product = 5, blind_code = TRUE)
design

create_template(design, attribute = c("Sweetness", "Mint", "Green", "Herbal"))
}
