% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_prefmap.R
\name{perform_prefmap}
\alias{perform_prefmap}
\title{Perform External Preference Mapping}
\usage{
perform_prefmap(data)
}
\arguments{
\item{data}{a sensory table}
}
\description{
Perform external preference mapping on sensory table.
}
