% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare sensory experiment}
\usage{
prepare(n_panelist, product, blind_code = FALSE, seed = NULL)
}
\arguments{
\item{n_panelist}{number of panelist}

\item{product}{a numeric value of number or product or vector of product names}

\item{blind_code}{wheteher to generate random three digit number for labeling}

\item{seed}{an integer for anchoring randomisation}
}
\value{
a dataframe of sensory design with panelist column and order of product presentation columns
}
\description{
Generate a design of presentation order based on Williams design.
}
\examples{
prepare(n_panelist = 30, product = letters[1:14], blind_code = FALSE)
prepare(n_panelist = 30, product = letters[1:14], blind_code = TRUE)

design <- prepare(n_panelist = 20, product = 5, blind_code = TRUE)
design
}
