% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare data}
\usage{
prepare(.data, type)
}
\arguments{
\item{.data}{result from get_stats}

\item{type}{type of original data, valid values: "string" and "number"}
}
\value{
a tibble
}
\description{
Prepare result from local test for global test
}
